Remove input encoding hack from manual.
In a previous commit (1a278d6f5ea3a2b2234464af4e908bf130a59d10) we have
converted the c file to utf-8 so it's no longer necessary to change the
input encoding.

I have decided to leave the commands in a comment until we decide to
modify the text to reflect this change.