Allow utf-8 in bibliography
This means that we have to switch the bibliography backend to biblatex
and use biber to compile the files.
Note that this commit doesn't compile with the makefiles because of
backend or packages issues, but it seems to be a utf-8 issue that is
solved when we use the rest of the changes of this pull request.