Make listings work with utf-8 encoded source files.
We solve it by defining a list of characters and it corresponding
accented translation, like old-style latex. For example á is {á}{{\'a}}

This is because the way the inputenc package works with non-ASCII
UTF-8-encoded characters (by making the first byte active and then
reading the following ones as arguments) is fundamentally incompatible
with the way the listing package works, which reads each byte
individually and expects it to be an individual character.
See https://tex.stackexchange.com/questions/24528/having-problems-with-listings-and-utf-8-can-it-be-fixed