Fix color of header in cover.
In 01b0e11c840645ab1877f5affb0cf35d6c6afe22 we broke the color of the
header because we made the opacity of the heading of the cover too low.
This commit solve it by introducing a redundant value of 1 that solve
the bug as it triggers the normal opacity check in xelatex.