Fix color of header in cover.
In 442e6922d911bc9b757f0d696cd0746f6f482ee6 we broke the color of the
header because we made the opacity of the heading of the cover too low.
This commit solve it by introducing a redundant value of 1 that solve
the bug as it triggers the normal opacity check in xelatex.