Add an option to use makenoidxglossaries.
This doesn't require an external program in contrast with makeglossaries,
and it works in overleaf but it may have problems if you use it in an
non utf8 context.

This is not activated by default because it non-recommended by the
package author, and we don't want to break an existing workflow.